コメント設定を完全に設定ダイアログに移動
